+++ src/skip_or_throw.c
@@ -2,7 +2,7 @@
 #include "throw.h"
 #include "skip_or_throw.h"
 
-void skip_or_throw(const size_t bytes)
+void skip_or_throw(const int bytes)
 {
   if (fseek(stdin, bytes, SEEK_CUR))
   {
