+++ src/skip_or_throw.h
@@ -2,13 +2,11 @@
 
 #define SKIP_OR_THROW_H
 
-#include <stddef.h>
-
 /**
  * Skips a specified number of bytes in standard input, halting execution should
  * the operation fail.
  * @param bytes The number of bytes to skip.
  */
-void skip_or_throw(const size_t bytes);
+void skip_or_throw(const int bytes);
 
 #endif
